+++ Assets/_Project/NPC/Scripts/AI/NPCState.cs
@@ -1,0 +1,12 @@
+///<summary>
+///Represents the high-level decision state of na NPC
+/// </summary>
+public enum NPCState
+{
+    Idle,
+    GoingToDestination,
+    Working,
+    Shopping,
+    GoingHome,
+    Fleeing
+}